\name{maize}
\alias{maize}
\docType{data}
\title{
Maize data
}
\description{
This is a dataset consisting of maize yields, environmental and management
variables measured across 4 groups. The groups pertain to different seasons
(2010 and 2013) for
different locations in Ethiopia (Pawe and Adami Tullu).
}
\usage{data("maize")}
\format{
  The format is:
List of 4
}
\details{
Contains a subset of data used in the Hermes et al. (2022) paper, which is
a subset
of data used in the Vasco Silva et al. (forthcoming) paper.
}
\source{
1. Hermes, S., van Heerwaarden, J., and Behrouzi, P. (2022). Copula graphical
models for heterogeneous mixed data. arXiv preprint, arXiv:2210.13140. \cr
}
\references{
1. Hermes, S., van Heerwaarden, J., and Behrouzi, P. (2022). Copula graphical
models for heterogeneous mixed data. arXiv preprint, arXiv:2210.13140. \cr
2. Vasco Silva, J., J. van Heerwaarden, R. Pytrik, A. G. Laborte, K. Tesfaye,
and M. K. van Ittersum (forthcoming). Big data, small explanatory power?
lessons learnt with random forest predictive modeling of crop yield in
contrasting farming systems. \cr
}
\examples{
data(maize)
}
\keyword{datasets}
